refactor and add license
